Add a margin class to the block edit wrapper when a margin is set
This forgoes the data attribute and just adds a class.

It should be noted that things might get a little wonky for dynamic
blocks that use ServerSideRender to display output in the editor.
Since discovering the `get_block_wrapper_attributes()` function,
I've been including it in for every block that fits the aforementioned
critera - in which case, the margin class is applied both the block
output and the block wrapper. However, thanks to margin collapsing,
this doesn't result in a visual doubling of the margin - at least not
that I saw during testing. Fingers crossed that continues to be the case :)